+++ Src/FluentAssertions/Types/TypeSelector.cs
@@ -202,6 +202,24 @@ public class TypeSelector : IEnumerable<Type>
     }
 
     /// <summary>
+    /// Filters and returns the types that are sealed
+    /// </summary>
+    public TypeSelector ThatAreSealed()
+    {
+        types = types.Where(t => t.IsSealed).ToList();
+        return this;
+    }
+
+    /// <summary>
+    /// Filters and returns the types that are not sealed
+    /// </summary>
+    public TypeSelector ThatAreNotSealed()
+    {
+        types = types.Where(t => !t.IsSealed).ToList();
+        return this;
+    }
+
+    /// <summary>
     /// Determines whether the type is static
     /// </summary>
     public TypeSelector ThatAreStatic()