+++ Tests/FluentAssertions.Specs/Types/TypeSelectorSpecs.cs
@@ -9,6 +9,7 @@ using Internal.Main.Test;
 using Internal.NotOnlyClasses.Test;
 using Internal.Other.Test;
 using Internal.Other.Test.Common;
+using Internal.SealedAndNotSealedClasses.Test;
 using Internal.StaticAndNonStaticClasses.Test;
 using Internal.UnwrapSelectorTestTypes.Test;
 using Xunit;
@@ -530,6 +531,40 @@ namespace FluentAssertions.Specs.Types
         }
 
         [Fact]
+        public void When_selecting_types_that_are_sealed_classes_it_should_return_the_correct_types()
+        {
+            // Arrange
+            Assembly assembly = typeof(SealedClass).GetTypeInfo().Assembly;
+
+            // Act
+            IEnumerable<Type> types = AllTypes.From(assembly)
+                .ThatAreInNamespace("Internal.SealedAndNotSealedClasses.Test")
+                .ThatAreSealed();
+
+            // Assert
+            types.Should()
+                .ContainSingle()
+                .Which.Should().Be(typeof(SealedClass));
+        }
+
+        [Fact]
+        public void When_selecting_types_that_are_not_sealed_classes_it_should_return_the_correct_types()
+        {
+            // Arrange
+            Assembly assembly = typeof(NotSealedClass).GetTypeInfo().Assembly;
+
+            // Act
+            IEnumerable<Type> types = AllTypes.From(assembly)
+                .ThatAreInNamespace("Internal.SealedAndNotSealedClasses.Test")
+                .ThatAreNotSealed();
+
+            // Assert
+            types.Should()
+                .ContainSingle()
+                .Which.Should().Be(typeof(NotSealedClass));
+        }
+
+        [Fact]
         public void When_selecting_types_that_are_static_classes_it_should_return_the_correct_types()
         {
             // Arrange
@@ -778,6 +813,17 @@ namespace Internal.UnwrapSelectorTestTypes.Test
         IEnumerator IEnumerable.GetEnumerator() => ((IEnumerable)integers).GetEnumerator();
 
         IEnumerator<string> IEnumerable<string>.GetEnumerator() => strings.GetEnumerator();
+    }
+}
+
+namespace Internal.SealedAndNotSealedClasses.Test
+{
+    internal sealed class SealedClass
+    {
+    }
+
+    internal class NotSealedClass
+    {
     }
 }